🔧 스터디룸 파일 구조 변경
스터디룸 게시글, 스터디룸 트러블슈팅 게시글, 스터디룸 정보나눔 게시판 각각의 엔티티들을 패키지로 분류
