+++ src/main/java/gaji/service/domain/roomBoard/converter/RoomPostConverter.java
@@ -1,10 +1,10 @@
 package gaji.service.domain.roomBoard.converter;
 
-import gaji.service.domain.roomBoard.entity.RoomInfoPost;
-import gaji.service.domain.roomBoard.entity.common.RoomBoard;
-import gaji.service.domain.roomBoard.entity.RoomPost;
-import gaji.service.domain.roomBoard.entity.RoomTroublePost;
-import gaji.service.domain.roomBoard.entity.TroublePostComment;
+import gaji.service.domain.roomBoard.entity.RoomInfo.RoomInfoPost;
+import gaji.service.domain.roomBoard.entity.RoomBoard;
+import gaji.service.domain.roomBoard.entity.RoomPost.RoomPost;
+import gaji.service.domain.roomBoard.entity.RoomTrouble.RoomTroublePost;
+import gaji.service.domain.roomBoard.entity.RoomTrouble.TroublePostComment;
 import gaji.service.domain.roomBoard.web.dto.RoomPostRequestDto;
 import gaji.service.domain.roomBoard.web.dto.RoomPostResponseDto;
 import gaji.service.domain.studyMate.entity.StudyMate;
