+++ src/main/java/gaji/service/domain/roomBoard/entity/RoomInfo/RoomInfoPost.java
@@ -1,14 +1,11 @@
-package gaji.service.domain.roomBoard.entity;
+package gaji.service.domain.roomBoard.entity.RoomInfo;
 
-import gaji.service.domain.roomBoard.entity.common.*;
+import gaji.service.domain.roomBoard.entity.RoomBoard;
 import gaji.service.domain.studyMate.entity.StudyMate;
-import gaji.service.domain.user.entity.User;
 import jakarta.persistence.*;
 import lombok.*;
 
 import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.List;
 
 @Entity
 @Getter
