+++ src/main/java/gaji/service/domain/roomBoard/entity/RoomPost/RoomComment.java
@@ -1,7 +1,6 @@
-package gaji.service.domain.roomBoard.entity.common;
+package gaji.service.domain.roomBoard.entity.RoomPost;
 
 import gaji.service.domain.enums.UserAlarmTypeEnum;
-import gaji.service.domain.roomBoard.entity.RoomPost;
 import gaji.service.domain.user.entity.User;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
