+++ src/main/java/gaji/service/domain/roomBoard/entity/RoomPost/RoomCommentLikes.java
@@ -1,4 +1,4 @@
-package gaji.service.domain.roomBoard.entity.common;
+package gaji.service.domain.roomBoard.entity.RoomPost;
 
 import gaji.service.domain.user.entity.User;
 import jakarta.persistence.*;
