+++ src/main/java/gaji/service/domain/roomBoard/entity/RoomPost/RoomPost.java
@@ -1,6 +1,6 @@
-package gaji.service.domain.roomBoard.entity;
+package gaji.service.domain.roomBoard.entity.RoomPost;
 
-import gaji.service.domain.roomBoard.entity.common.*;
+import gaji.service.domain.roomBoard.entity.RoomBoard;
 import gaji.service.domain.user.entity.User;
 import jakarta.persistence.*;
 import lombok.*;
