+++ src/main/java/gaji/service/domain/roomBoard/entity/RoomPost/RoomPostLikes.java
@@ -1,6 +1,5 @@
-package gaji.service.domain.roomBoard.entity.common;
+package gaji.service.domain.roomBoard.entity.RoomPost;
 
-import gaji.service.domain.roomBoard.entity.RoomPost;
 import gaji.service.domain.user.entity.User;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
@@ -10,19 +9,16 @@ import lombok.NoArgsConstructor;
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class RoomPostFile {
+public class RoomPostLikes {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "post_id")
-    private RoomPost roomPost;
-
-    @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id")
     private User user;
 
-    private String path;
-
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "post_id")
+    private RoomPost roomPost;
 }
