+++ src/main/java/gaji/service/domain/roomBoard/entity/RoomPost/RoomPostReport.java
@@ -1,7 +1,7 @@
-package gaji.service.domain.roomBoard.entity.common;
+package gaji.service.domain.roomBoard.entity.RoomPost;
 
-import gaji.service.domain.roomBoard.entity.RoomPost;
 import gaji.service.domain.user.entity.User;
+import gaji.service.domain.enums.ReportPostTypeEnum;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
@@ -10,8 +10,7 @@ import lombok.NoArgsConstructor;
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class RoomPostBookmark {
-
+public class RoomPostReport {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
@@ -21,6 +20,11 @@ public class RoomPostBookmark {
     private User user;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "room_id")
+    @JoinColumn(name = "post_id")
     private RoomPost roomPost;
+
+    @Enumerated(EnumType.STRING)
+    private ReportPostTypeEnum reportPostTypeEnum;
+
+    private String description;
 }
