+++ src/main/java/gaji/service/domain/roomBoard/repository/RoomInfoPostRepository.java
@@ -1,6 +1,6 @@
 package gaji.service.domain.roomBoard.repository;
 
-import gaji.service.domain.roomBoard.entity.RoomInfoPost;
+import gaji.service.domain.roomBoard.entity.RoomInfo.RoomInfoPost;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
