+++ src/main/java/gaji/service/domain/roomBoard/repository/RoomPostCommentRepository.java
@@ -1,7 +1,7 @@
 package gaji.service.domain.roomBoard.repository;
 
 
-import gaji.service.domain.roomBoard.entity.TroublePostComment;
+import gaji.service.domain.roomBoard.entity.RoomTrouble.TroublePostComment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RoomPostCommentRepository extends JpaRepository<TroublePostComment,Long> {
