+++ src/main/java/gaji/service/domain/roomBoard/repository/RoomPostRepository.java
@@ -1,6 +1,6 @@
 package gaji.service.domain.roomBoard.repository;
 
-import gaji.service.domain.roomBoard.entity.RoomPost;
+import gaji.service.domain.roomBoard.entity.RoomPost.RoomPost;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
