+++ src/main/java/gaji/service/domain/roomBoard/repository/RoomTroublePostBookmarkRepository.java
@@ -1,7 +1,7 @@
 package gaji.service.domain.roomBoard.repository;
 
-import gaji.service.domain.roomBoard.entity.RoomTroublePost;
-import gaji.service.domain.roomBoard.entity.RoomTroublePostBookmark;
+import gaji.service.domain.roomBoard.entity.RoomTrouble.RoomTroublePost;
+import gaji.service.domain.roomBoard.entity.RoomTrouble.RoomTroublePostBookmark;
 import gaji.service.domain.studyMate.entity.StudyMate;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
