+++ src/main/java/gaji/service/domain/roomBoard/repository/RoomTroublePostRepository.java
@@ -1,6 +1,6 @@
 package gaji.service.domain.roomBoard.repository;
 
-import gaji.service.domain.roomBoard.entity.RoomTroublePost;
+import gaji.service.domain.roomBoard.entity.RoomTrouble.RoomTroublePost;
 import gaji.service.domain.roomBoard.web.dto.RoomPostResponseDto;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
