+++ src/main/java/gaji/service/domain/roomBoard/repository/TroublePostCommentRepository.java
@@ -1,6 +1,6 @@
 package gaji.service.domain.roomBoard.repository;
 
-import gaji.service.domain.roomBoard.entity.TroublePostComment;
+import gaji.service.domain.roomBoard.entity.RoomTrouble.TroublePostComment;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
