+++ src/main/java/gaji/service/domain/roomBoard/service/RoomInfoPostCommandServiceImpl.java
@@ -4,8 +4,8 @@ import gaji.service.domain.enums.RoomPostType;
 import gaji.service.domain.room.entity.Room;
 import gaji.service.domain.room.service.RoomQueryService;
 import gaji.service.domain.roomBoard.converter.RoomPostConverter;
-import gaji.service.domain.roomBoard.entity.RoomInfoPost;
-import gaji.service.domain.roomBoard.entity.common.RoomBoard;
+import gaji.service.domain.roomBoard.entity.RoomInfo.RoomInfoPost;
+import gaji.service.domain.roomBoard.entity.RoomBoard;
 import gaji.service.domain.roomBoard.repository.RoomBoardRepository;
 import gaji.service.domain.roomBoard.repository.RoomInfoPostRepository;
 import gaji.service.domain.roomBoard.repository.RoomTroublePostRepository;
