+++ src/main/java/gaji/service/domain/roomBoard/service/RoomPostCommandService.java
@@ -1,6 +1,6 @@
 package gaji.service.domain.roomBoard.service;
 
-import gaji.service.domain.roomBoard.entity.RoomPost;
+import gaji.service.domain.roomBoard.entity.RoomPost.RoomPost;
 import gaji.service.domain.roomBoard.web.dto.RoomPostRequestDto;
 
 public interface RoomPostCommandService {
