+++ src/main/java/gaji/service/domain/roomBoard/service/RoomTroublePostCommandServiceImpl.java
@@ -5,8 +5,11 @@ import gaji.service.domain.room.entity.Room;
 import gaji.service.domain.room.service.RoomQueryService;
 import gaji.service.domain.roomBoard.code.RoomPostErrorStatus;
 import gaji.service.domain.roomBoard.converter.RoomPostConverter;
-import gaji.service.domain.roomBoard.entity.*;
-import gaji.service.domain.roomBoard.entity.common.RoomBoard;
+import gaji.service.domain.roomBoard.entity.RoomTrouble.RoomTroublePost;
+import gaji.service.domain.roomBoard.entity.RoomTrouble.RoomTroublePostBookmark;
+import gaji.service.domain.roomBoard.entity.RoomTrouble.RoomTroublePostLike;
+import gaji.service.domain.roomBoard.entity.RoomTrouble.TroublePostComment;
+import gaji.service.domain.roomBoard.entity.RoomBoard;
 import gaji.service.domain.roomBoard.repository.*;
 import gaji.service.domain.roomBoard.web.dto.RoomPostRequestDto;
 import gaji.service.domain.roomBoard.web.dto.RoomPostResponseDto;
