+++ src/main/java/gaji/service/domain/studyMate/entity/StudyMate.java
@@ -3,10 +3,10 @@ package gaji.service.domain.studyMate.entity;
 import gaji.service.domain.enums.Role;
 import gaji.service.domain.room.entity.Room;
 import gaji.service.domain.room.entity.RoomNotice;
-import gaji.service.domain.roomBoard.entity.RoomTroublePostBookmark;
-import gaji.service.domain.roomBoard.entity.RoomTroublePostLike;
-import gaji.service.domain.roomBoard.entity.RoomInfoPost;
-import gaji.service.domain.roomBoard.entity.RoomTroublePost;
+import gaji.service.domain.roomBoard.entity.RoomTrouble.RoomTroublePostBookmark;
+import gaji.service.domain.roomBoard.entity.RoomTrouble.RoomTroublePostLike;
+import gaji.service.domain.roomBoard.entity.RoomInfo.RoomInfoPost;
+import gaji.service.domain.roomBoard.entity.RoomTrouble.RoomTroublePost;
 import gaji.service.domain.user.entity.User;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
